added classes completed and classes not edited table
