+++ Managementsystem_Classconferences/Managementsystem_Classconferences/Pages/Diplomarbeit/Moderator/Models/Page_Content.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Managementsystem_Classconferences.Pages.Diplomarbeit.Moderator.Models
+{
+    public class Page_Content
+    {
+        public string Teachers { get; set; }
+        public string Intersections { get; set; }
+        public string Classname { get; set; }
+        public string Buttontext { get; set; }
+    }
+}
